refactor(agent-runtime): extract compaction summarization to consumer-provided CompactionStrategy
The runtime's compactMessages no longer hardcodes the summarization prompt
or LLM call. Consumers inject a CompactionStrategy (with the summarization
logic). If no strategy is provided, compaction is silently skipped.
