+++ packages/agent-runtime/src/index.ts
@@ -3,4 +3,4 @@ export { AgentRuntime } from './AgentRuntime'
 export { buildPromptWithTurnContext } from './turnContext'
 export type { TurnContextOptions } from './turnContext'
 export type { RuntimeStartInput } from './types'
-export type { AgentRuntimeConfig, AIProviderFactory, IAgentEventEmitter, IAgentPathProvider, IAIProvider, IAIStreamChunk, IConversationQuery, ILogger, IModelResolver, LoopMessage, RuntimeToolDefinition, ToolProvider } from '@ant-chat/shared'
+export type { AgentRuntimeConfig, AIProviderFactory, CompactionStrategy, IAgentEventEmitter, IAgentPathProvider, IAIProvider, IAIStreamChunk, IConversationQuery, ILogger, IModelResolver, LoopMessage, RuntimeToolDefinition, ToolProvider } from '@ant-chat/shared'
